Renaming plus to resistance
